backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.7.4
pydantic-settings==2.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
redis==5.0.7
celery==5.4.0
httpx==0.27.0
structlog==24.2.0
anthropic==0.28.0
numpy==1.26.4
pandas==2.2.2
ta==0.11.0
scikit-learn==1.5.0
hmmlearn==0.3.2
websockets==12.0
orjson==3.10.5
